#87 - refactor: 댓글 Controller 단일 책임 원칙 적용
