--- src/main/java/com/gxdxx/instagram/domain/comment/api/CommentController.java
@@ -1,94 +1,0 @@
-package com.gxdxx.instagram.domain.comment.api;
-
-import com.gxdxx.instagram.domain.comment.dto.request.CommentRegisterRequest;
-import com.gxdxx.instagram.domain.comment.dto.request.CommentUpdateRequest;
-import com.gxdxx.instagram.domain.comment.dto.response.CommentRegisterResponse;
-import com.gxdxx.instagram.domain.comment.dto.response.CommentUpdateResponse;
-import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
-import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
-import com.gxdxx.instagram.global.error.InvalidRequestException;
-import com.gxdxx.instagram.domain.comment.application.CommentCreateService;
-import com.gxdxx.instagram.domain.comment.application.CommentDeleteService;
-import com.gxdxx.instagram.domain.comment.application.CommentUpdateService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-
-import java.security.Principal;
-
-@Tag(name = "comments", description = "댓글 API")
-@RequiredArgsConstructor
-@RequestMapping("/api/comments")
-@RestController
-public class CommentController {
-
-    private final CommentCreateService commentCreateService;
-    private final CommentUpdateService commentUpdateService;
-    private final CommentDeleteService commentDeleteService;
-
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "댓글 작성 성공",
-                    content = @Content(schema = @Schema(implementation = CommentRegisterResponse.class))),
-            @ApiResponse(responseCode = "400", description = "잘못된 요청",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @Operation(summary = "댓글 작성 메소드", description = "댓글 작성 메소드입니다.")
-    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public CommentRegisterResponse createComment(
-            @RequestBody @Valid CommentRegisterRequest request,
-            BindingResult bindingResult,
-            Principal principal
-    ) {
-        validateRequest(bindingResult);
-        return commentCreateService.createComment(request, principal.getName());
-    }
-
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "댓글 수정 성공",
-                    content = @Content(schema = @Schema(implementation = CommentUpdateResponse.class))),
-            @ApiResponse(responseCode = "400", description = "잘못된 요청",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @Operation(summary = "댓글 수정 메소드", description = "댓글 수정 메소드입니다.")
-    @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public CommentUpdateResponse updateComment(
-            @RequestBody @Valid CommentUpdateRequest request,
-            BindingResult bindingResult,
-            Principal principal
-    ) {
-        validateRequest(bindingResult);
-        return commentUpdateService.updateComment(request, principal.getName());
-    }
-
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "204", description = "댓글 삭제 성공",
-                    content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
-            @ApiResponse(responseCode = "400", description = "잘못된 요청",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-    })
-    @Operation(summary = "댓글 삭제 메소드", description = "댓글 삭제 메소드입니다.")
-    @DeleteMapping(value = "/{id}", consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public SuccessResponse deleteComment(
-            @Parameter(name = "id", description = "댓글의 id", in = ParameterIn.PATH) @PathVariable("id") Long id,
-            Principal principal
-    ) {
-        return commentDeleteService.deleteComment(id, principal.getName());
-    }
-
-    private void validateRequest(BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            throw new InvalidRequestException();
-        }
-    }
-
-}
