+++ src/main/java/com/gxdxx/instagram/domain/comment/api/CommentDeleteController.java
@@ -1,0 +1,56 @@
+package com.gxdxx.instagram.domain.comment.api;
+
+import com.gxdxx.instagram.domain.comment.application.CommentUpdateService;
+import com.gxdxx.instagram.domain.comment.dto.request.CommentUpdateRequest;
+import com.gxdxx.instagram.domain.comment.dto.response.CommentUpdateResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
+
+@Tag(name = "comments", description = "댓글 API")
+@RequiredArgsConstructor
+@RequestMapping("/api/comments")
+@RestController
+public class CommentDeleteController {
+
+    private final CommentUpdateService commentUpdateService;
+
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "댓글 수정 성공",
+                    content = @Content(schema = @Schema(implementation = CommentUpdateResponse.class))),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청",
+                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    })
+    @Operation(summary = "댓글 수정 메소드", description = "댓글 수정 메소드입니다.")
+    @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public CommentUpdateResponse updateComment(
+            @RequestBody @Valid CommentUpdateRequest request,
+            BindingResult bindingResult,
+            Principal principal
+    ) {
+        validateRequest(bindingResult);
+        return commentUpdateService.updateComment(request, principal.getName());
+    }
+
+    private void validateRequest(BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new InvalidRequestException();
+        }
+    }
+
+}
